requirements.txt
django
paho-mqtt
pyyaml
sqlalchemy
openpyxl
django-simpleui
channels
pandas
whitenoise
django-cors-headers
gunicorn>=20.1.0
fiona==1.9.4
geopandas==0.13.2
networkx
